+++ src/main/java/entity/AIPortAccount.java
@@ -1,0 +1,53 @@
+package entity;
+
+public class AIPortAccount {
+    public long id;
+    public String account;
+    public String password;
+    public int status;
+    public String keySeed;
+
+    // Constructor
+    public AIPortAccount() {}
+
+    // Getters and setters
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getAccount() {
+        return account;
+    }
+
+    public void setAccount(String account) {
+        this.account = account;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public int getStatus() {
+        return status;
+    }
+
+    public void setStatus(int status) {
+        this.status = status;
+    }
+
+    public String getKeySeed() {
+        return keySeed;
+    }
+
+    public void setKeySeed(String keySeed) {
+        this.keySeed = keySeed;
+    }
+}