+++ src/main/java/entity/AIPortTeamMember.java
@@ -1,0 +1,53 @@
+package entity;
+
+public class AIPortTeamMember {
+    public long teamId;
+    public long memberId;
+    public short status;
+    public long created;
+    public long expirationDate;
+
+    // Constructor
+    public AIPortTeamMember() {}
+
+    // Getters and setters
+    public long getTeamId() {
+        return teamId;
+    }
+
+    public void setTeamId(long teamId) {
+        this.teamId = teamId;
+    }
+
+    public long getMemberId() {
+        return memberId;
+    }
+
+    public void setMemberId(long memberId) {
+        this.memberId = memberId;
+    }
+
+    public short getStatus() {
+        return status;
+    }
+
+    public void setStatus(short status) {
+        this.status = status;
+    }
+
+    public long getCreated() {
+        return created;
+    }
+
+    public void setCreated(long created) {
+        this.created = created;
+    }
+
+    public long getExpirationDate() {
+        return expirationDate;
+    }
+
+    public void setExpirationDate(long expirationDate) {
+        this.expirationDate = expirationDate;
+    }
+}